Moved TcpClient project to samples/Lab2
