+++ samples/Lab2/NetworkProgramming.Lab2/AbstractClient.cs
