+++ samples/Lab2/NetworkProgramming.Lab2/Client.cs
