+++ samples/Lab2/NetworkProgramming.Lab2/ConsoleGui.cs
