+++ samples/Lab2/NetworkProgramming.Lab2/Logger.cs
