+++ samples/Lab2/NetworkProgramming.Lab2/NetworkProgramming.Lab2.csproj
