+++ samples/Lab2/NetworkProgramming.Lab2/SocketExtensions.cs
@@ -1,0 +1,15 @@
+﻿using System.Net.Sockets;
+using System.Reflection;
+
+namespace NetworkProgramming.Lab1
+{
+	public static class SocketExtensions
+	{
+		public static bool IsDisposed(this Socket socket)
+		{
+			const BindingFlags flags = BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.GetProperty;
+			var field = socket.GetType().GetProperty("CleanedUp", flags);
+			return (bool) (field?.GetValue(socket, null) ?? false);
+		}
+	}
+}
